+++ lib/pages/setting/device_settings/shelf_management/shelf_management_light/controller.dart
@@ -120,6 +120,10 @@ class ShelfManagementLightController extends GetxController {
 
   // 删除
   void delete() async {
+    if (currentSection.value.isEmpty) {
+      PopupMessage.showWarningInfoBar('请选择要删除的节点');
+      return;
+    }
     var res = await CommonApi.deleteSection({
       "params": [
         {
